refactor(auth): resserre le socle MFA après les gates qualité
- type nommé pour les paramètres d'autorisation de step-up (corrige un
  typecheck sous noUncheckedIndexedAccess)
- contexte de requête (IP, user-agent) sur la ligne d'audit du passage MFA,
  comme les autres événements auth du fichier
- ADMIN_EMAILS dans le mock env partagé plutôt qu'un override local qui
  recopiait sept valeurs déjà maintenues dans src/test/setup.ts
- tests du marqueur MFA sortis dans leur propre fichier, comme les autres
  concerns du config auth
- commentaires ramenés au format du projet, référence au ticket retirée
